+++ src/valueaccess.rs
@@ -2,7 +2,7 @@ use crate::xmlvalue::{Comment, Element, ProcessingInstruction, Text, Value, Valu
 use crate::xotdata::{Node, Xot};
 
 /// ## Value and type access
-impl Xot {
+impl<'a> Xot<'a> {
     /// Access to the XML value for this node.
     ///
     /// ```rust